pyproject.toml
[tool.poetry]
name = "grpc-asyncio-example"
version = "0.1.0"
description = ""
authors = ["so1n <qaz6803609@163.com>"]
readme = "README.md"
packages = [{include = "grpc_asyncio_example"}]

[tool.poetry.dependencies]
python = "^3.7"
grpcio = "1.46.1"
grpcio-tools = "^1.46.1"
betterproto = {version = "2.0.0b5", extras = ["compiler"]}
starlette = "^0.22.0"
uvicorn = "^0.20.0"
py-spy = "^0.3.14"

[tool.poetry.dev-dependencies]
pre-commit = "^2.14.0"
mypy-protobuf = "^2.6"

[tool.poetry.group.dev.dependencies]
uvloop = "^0.17.0"
gevent = "^22.10.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
